requirements/base.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Async & Tasks
celery==5.3.4
redis==5.0.1
daphne==4.0.0
channels-redis==4.2.0
django-redis==5.4.0

# Environnement & Config
python-dotenv==1.0.0
django-environ==0.11.2

# API & Serialization
drf-spectacular==0.26.5
django-filter==23.3
django-cors-headers==4.3.0

# Security & Rate Limiting
django-ratelimit==4.1.0
django-csp==3.7

# Payments
stripe==7.4.0

# File Handling
Pillow==10.1.0

# Phone & SMS
phonenumbers==8.13.25
twilio==8.10.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

channels==4.0.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
pytest-asyncio==0.21.1